[tools] Split powertools out into expotools
This is a subset of powertools that we need in this repo to do things like publish dev versions of home. This commit copies most of powertools with the following changes:

- Paths are updated for the Expo repo
- Removed commands that belong more in powertools
- Changed uses of `exp` to use `expo` (Expo CLI) instead
- Removed old bootstrap/expu/yarnu scripts
- Removed fixtures for SDK versions we don't include in the client anymore
- Converted from Flow to TypeScript
- Moved sensitive environment variables out of the scripts and into an encrypted file

Note: the commands to published (unversioned/versioned) test suite just use whatever version is in app.json.

Tested the publish-dev-home command and verified that after building the iOS app with the `PUBLISHED` manifest (twice, to flush caches) Home loaded fine.